requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pandas
openpyxl
pydantic[email]
python-dotenv
pydantic-settings
